# data/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "System Initiative API",
    "description": "The API Server for interacting with a System Initiative workspace",
    "version": "1.0.0"
  },
  "servers": [
    {
      "url": "/",
      "description": "This API Server"
    }
  ],
  "paths": {
...
    "/v1/w/{workspace_id}/change-sets/{change_set_id}/secrets/{secret_id}": {
      "put": {
        "tags": [
          "secrets"
        ],
        "operationId": "update_secret",
        "parameters": [
          {
            "name": "workspace_id",
            "in": "path",
            "description": "Workspace identifier",
            "required": true
          },
          {
            "name": "change_set_id",
            "in": "path",
            "description": "Change set identifier",
            "required": true
          },
          {
            "name": "secret_id",
            "in": "path",
            "description": "Secret identifier",
            "required": true
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateSecretV1Request"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Secret updated successfully",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/UpdateSecretV1Response"
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized - Invalid or missing token"
          },
          "404": {
            "description": "Secret not found"
          },
          "500": {
            "description": "Internal server error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            }
          }
        }
      },
      "delete": {
        "tags": [
          "secrets"
        ],
        "operationId": "delete_secret",
        "parameters": [
          {
            "name": "workspace_id",
            "in": "path",
            "description": "Workspace identifier",
            "required": true
          },
          {
            "name": "change_set_id",
            "in": "path",
            "description": "Change set identifier",
            "required": true
          },
          {
            "name": "secret_id",
            "in": "path",
            "description": "Secret identifier",
            "required": true
          }
        ],
        "responses": {
          "200": {
            "description": "Secret deleted successfully",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DeleteSecretV1Response"
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized - Invalid or missing token"
          },
          "404": {
            "description": "Secret not found"
          },
          "500": {
            "description": "Internal server error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "ApiError": {
        "type": "object",
        "description": "Standard error response format for v1 API",
        "required": [
          "message",
          "statusCode"
        ],
        "properties": {
          "code": {
            "type": [
              "integer",
              "null"
            ],
            "format": "int32",
            "example": 4001
          },
          "message": {
            "type": "string",
            "example": "Invalid request data"
          },
          "statusCode": {
            "type": "integer",
            "format": "int32",
            "example": 422,
            "minimum": 0
          }
        },
        "example": {
          "message": "Invalid request data",
          "statusCode": 422,
          "code": 4001
        }
      },
...
      "DeleteSecretV1Response": {
        "type": "object",
        "required": [
          "success"
        ],
        "properties": {
          "success": {
            "type": "boolean"
          }
        }
      },
...
      "SecretV1": {
        "type": "object",
        "required": [
          "id",
          "name",
          "definition",
          "description"
        ],
        "properties": {
          "definition": {
            "type": "string",
            "example": "aws_credentials"
          },
          "description": {
            "type": "string",
            "example": "AWS credentials for production environment"
          },
          "id": {
            "type": "string",
            "example": "01HAXYZF3GC9CYA6ZVSM3E4YHH"
          },
          "name": {
            "type": "string",
            "example": "Production AWS Key"
          }
        }
      },
...
      "UpdateSecretV1Request": {
        "type": "object",
        "required": [
          "id",
          "name",
          "description"
        ],
        "properties": {
          "description": {
            "type": "string",
            "example": "Updated AWS Secret Key for EC2 access"
          },
          "id": {
            "type": "string",
            "example": "01HAXYZF3GC9CYA6ZVSM3E4YHH"
          },
          "name": {
            "type": "string",
            "example": "AWS Access Key"
          },
          "rawData": {
            "type": "object",
            "additionalProperties": {
              "type": "string"
            },
            "propertyNames": {
              "type": "string"
            },
            "example": {
              "access_key_id": "AKIAIOSFODNN7EXAMPLE",
              "secret_access_key": "wJalrXUtnFEMI/K7MDENG/bPxRfiCYEXAMPLEKEY",
              "region": "us-west-2",
              "default_output": "json"
            }
          }
        }
      },
      "UpdateSecretV1Response": {
        "type": "object",
        "required": [
          "secret"
        ],
        "properties": {
          "secret": {
            "$ref": "#/components/schemas/SecretV1"
          }
        }
      },
...
    }
  },
...
}